Add hebcal (C) as a submodule.
For now this links to my fork of hebcal, which has a bunch of
enhancements specifically geared towards being used in this way,
linked as a shared library and called via JNI. (Some of the changes
may date back to the time when I was running this code on Palm,
though I have tried to clean the worst of those.)

Years back I also added some extra zemanim support, which needs to be
rehashed a bit in light of extensive upstream changes to hebcal's handling
of time. With luck when I get around to that update I hope it may happen
to fix the end-of-year AM/PM swap bug.